+++ changes.md
@@ -10,6 +10,7 @@
 * Introduced cli commands
 * Added possibility to only use command line to generate Home Assistant Configuration
 * Application info added
+* Application info and description added to auto-generated Home Assistant configuration.
 
 ## v0.1.4 Bug fixed in python package
 * Bug in python package fixed